python-app/requirements.txt
PyQt5>=5.15.0
pynput>=1.7.6
Pillow>=9.0.0
pyautogui>=0.9.53
pyperclip>=1.8.2
pywin32>=305
